Changes made as per review comments
